docs: remove `(Optional)` from FAQ
Some ports in the Catppuccin namespace
haven't removed the (Optional) from the
FAQ heading even when FAQ questions 
have been included.

This commit reduces the chance of that
by removing it altogether. An empty FAQ
section would still get deleted and if 
FAQ questions are included, the maintainer
does not need to remove the "(Optional)"
from the heading anymore.